Reconciled the client by removing erroneous Auth0 template stuff. Merged navbar and LandingPage.
