+++ docker-compose.yaml
@@ -19,7 +19,7 @@ services:
     build: ./server
     
     ports:
-      - "1000:8080"
+      - "8081:8080"
     networks:
       - my_network
     environment:
@@ -31,7 +31,7 @@ services:
   frontend:
     build: ./client
     ports:
-      - "1001:3000"
+      - "3001:3000"
     networks:
       - my_network
     depends_on:
